agents.api/requirements.txt
fastapi>=0.115.8,<0.116.0
uvicorn>=0.34.0,<0.35.0
websockets>=14.2,<15.0
atomic-agents>=1.0.19,<2.0.0
openai>=1.63.0,<2.0.0
instructor>=1.7.2,<2.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.10.6,<3.0.0